homework_1/P3/requirements.txt
numpy 
pandas
networkx
matplotlib.pyplot